Extracted holochain_runtime crate
