+++ crates/holochain_runtime/src/happs/mod.rs
@@ -1,0 +1,2 @@
+pub mod install;
+pub mod update;
