+++ crates/holochain_runtime/src/launch.rs
@@ -9,22 +9,21 @@ use holochain::conductor::Conductor;
 use holochain_client::AdminWebsocket;
 
 use crate::{
-    filesystem::FileSystem,
-    launch::signal::{can_connect_to_signal_server, run_local_signal_service},
-    GossipArcClamp, HolochainPluginConfig, HolochainRuntime,
+    filesystem::FileSystem, launch::signal::{can_connect_to_signal_server, run_local_signal_service}, GossipArcClamp, HolochainRuntime, HolochainRuntimeConfig
 };
 
 mod mdns;
 mod signal;
+mod config;
 use mdns::spawn_mdns_bootstrap;
 
 // pub static RUNNING_HOLOCHAIN: RwLock<Option<RunningHolochainInfo>> = RwLock::const_new(None);
 
 /// Launch the holochain conductor in the background
-pub async fn launch_holochain_runtime(
+pub(crate) async fn launch_holochain_runtime(
     passphrase: BufRead,
-    config: HolochainPluginConfig,
-) -> crate::Result<HolochainRuntime> {
+    config: HolochainRuntimeConfig,
+) -> crate::error::Result<HolochainRuntime> {
     // let mut lock = RUNNING_HOLOCHAIN.write().await;
 
     // if let Some(info) = lock.to_owned() {
@@ -62,7 +61,7 @@ pub async fn launch_holochain_runtime(
         signal_urls.insert(0, local_signal_url.clone());
     }
 
-    let config = crate::config::conductor_config(
+    let config = config::conductor_config(
         &filesystem,
         admin_port,
         filesystem.keystore_dir().into(),
@@ -115,32 +114,6 @@ pub async fn wait_until_admin_ws_is_available(admin_port: u16) -> crate::Result<
     }
     Ok(())
 }
-
-// pub async fn wait_until_app_ws_is_available(app_port: u16) -> crate::Result<()> {
-//     let mut retry_count = 0;
-//     let _admin_ws = loop {
-//         if let Ok(ws) = AppWebsocket::connect(format!("localhost:{}", app_port))
-//             .await
-//             .map_err(|err| {
-//                 crate::Error::AdminWebsocketError(format!(
-//                     "Could not connect to the app interface: {}",
-//                     err
-//                 ))
-//             })
-//         {
-//             break ws;
-//         }
-//         async_std::task::sleep(Duration::from_millis(200)).await;
-
-//         retry_count += 1;
-//         if retry_count == 200 {
-//             return Err(crate::Error::AdminWebsocketError(
-//                 "Can't connect to holochain".to_string(),
-//             ));
-//         }
-//     };
-//     Ok(())
-// }
 
 // fn read_config(config_path: &std::path::Path) -> crate::Result<LairServerConfig> {
 //     let bytes = std::fs::read(config_path)?;
