+++ crates/holochain_runtime/src/lib.rs
@@ -1,0 +1,14 @@
+mod config;
+mod filesystem;
+mod launch;
+mod holochain_runtime;
+mod error;
+mod happs;
+mod lair_signer;
+
+pub use config::*;
+pub use error::*;
+pub use holochain_runtime::*;
+pub use lair_signer::*;
+pub use filesystem::*;
+pub use happs::update::UpdateHappError;
