+++ crates/tauri-plugin-holochain/src/commands/mod.rs
@@ -1,5 +1,3 @@
 pub mod get_runtime_info;
-pub mod install_web_app;
-pub mod list_apps;
 pub mod open_app;
 pub mod sign_zome_call;
