+++ flake.nix
@@ -480,4 +480,3 @@
       };
     };
 }
-
